+++ src/Products/MeetingLalouviere/config.py
@@ -35,9 +35,6 @@ PROJECTNAME = "MeetingLalouviere"
 DEFAULT_WRITE_FOLLOWUP_PERMISSION = "MeetingLalouviere: Write followUp"
 setDefaultRoles(DEFAULT_WRITE_FOLLOWUP_PERMISSION, ('Manager', 'MeetingManager'))
 
-DEFAULT_WRITE_NEEDED_FOLLOWUP_PERMISSION = "MeetingLalouviere: Write neededFollowUp"
-setDefaultRoles(DEFAULT_WRITE_NEEDED_FOLLOWUP_PERMISSION, ('Manager', 'MeetingManager'))
-
 DEFAULT_WRITE_PROVIDED_FOLLOWUP_PERMISSION = "MeetingLalouviere: Write providedFollowUp"
 setDefaultRoles(DEFAULT_WRITE_PROVIDED_FOLLOWUP_PERMISSION, ('Manager', 'MeetingManager', 'MeetingFollowUpWriter'))
 
@@ -232,7 +229,7 @@ LLO_APPLYED_COLLEGE_WFA = (
     'propose_to_budget_reviewer',
     'waiting_advices',
     'waiting_advices_proposing_group_send_back',
-    'item_validation_shortcuts,'
+    'item_validation_shortcuts'
 )
 
 LLO_APPLYED_COUNCIL_WFA = (